+++ src/FolkerKinzel.Strings/ReadOnlySpanPolyfillExtension_LastIndexOfAny .cs
@@ -1,9 +1,30 @@
+using FolkerKinzel.Strings.Intls;
+
 namespace FolkerKinzel.Strings;
 
 #if NET7_0 || NET6_0 || NET5_0 || NETCOREAPP3_1 || NETSTANDARD2_1 || NETSTANDARD2_0 || NET461
 
 public static partial class ReadOnlySpanPolyfillExtension
 {
+    /// <summary>
+    /// Searches for the last index of any of the specified Unicode characters.
+    /// </summary>
+    /// <param name="span">The read-only span to search.</param>
+    /// <param name="values">The set of characters to search for.</param>
+    /// <returns>The zero-based index of the last occurrence of one of the specified Unicode
+    /// characters in <paramref name="span" /> or -1 if none of these characters have been
+    /// found.</returns>
+    /// <exception cref="ArgumentNullException"><paramref name="values"/> is <c>null</c>.</exception>
+    public static int LastIndexOfAny(this ReadOnlySpan<char> span, SearchValues<char> values)
+    {
+        // Don't address MemoryExtensions here directly because the library method
+        // polyfills a bug in the nuget package System.Memory for .NET Framework and
+        // .NET Standard 2.0
+
+        _ArgumentNullException.ThrowIfNull(values, nameof(values));
+        return span.LastIndexOfAny(values.Span);
+    }
+
 #if NET461 || NETSTANDARD2_0
 
     /// <summary>Searches for the zero-based index of the last occurrence of one of the specified